TT-2: Add tests for the class MathUtil and minor refactoring
1) Create class MathUtil
2) Transfer method isNumeric to MathUtil
3) Add tests for method isNumeric
4) Minor refactoring
