+++ src/main/java/com/byndyusoft/util/exception/IncorrectFormatExpression.java
@@ -1,4 +1,4 @@
-package com.byndyusoft;
+package com.byndyusoft.util.exception;
 
 public class IncorrectFormatExpression extends RuntimeException {
     public IncorrectFormatExpression(String message) {
