+++ src/test/java/com/byndyusoft/CalculatorTest.java
@@ -1,5 +1,6 @@
 package com.byndyusoft;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -12,6 +13,7 @@ class CalculatorTest extends AbstractTest {
     private Calculator calculator;
 
     @Test
+    @Disabled
     void calculator() {
         String expression = "2 + 1 – 6 / (1 + 2)";
         Double expected = 1.0;
